Applying automated Edit/Remove Buttons
Applying automated Edit/Remove Buttons, the playerId is added to the onClick event, so that if the user wants to edit or remove a player, the player Id is provided to the function to perform its duties.

Also added commenting to make it more clear what each function is used for.