Edit Player Functionality
Adding functionality to edit a player. A pop up box is displayed to the user after clicking 'edit' next on the row associated with a player. The values for the player are retrieved from the array and autopopulated on the screen.

After clicking save, the associated players information will be changed, and reuploaded to local storage